+++ app/src/components/common/appInfo/AppInfoLauncher.tsx
@@ -1,7 +1,8 @@
 import React, { useState } from "react";
-import { ActionFactory, IconButtonWrapper } from "../IconButtonWrapper";
+import { IconButtonWrapper } from "../actions/IconButtonWrapper";
 import { AppInfo } from "./AppInfo";
 import { DialogWrapper } from "../../layout/dialogs/DialogWrapper";
+import { ActionFactory } from "../actions/ActionFactory";
 
 export const AppInfoLauncher: React.FC = () => {
   const [showInfo, setShowInfo] = useState(false);
