+++ app/src/components/details/filters/DateFilters.tsx
@@ -6,7 +6,7 @@ import {
   createDateConditions,
   createNextDateConditions,
 } from "./createDateConditions";
-import { IconButtonWrapper } from "../../common/IconButtonWrapper";
+import { IconButtonWrapper } from "../../common/actions/IconButtonWrapper";
 import { ChevronLeft, ChevronRight } from "@mui/icons-material";
 import { DateRangeSelector } from "./DateRangeSelector";
