+++ app/src/components/overview/MetricHeaderActions.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from "react";
 import { IMetric } from "../../serverApi/IMetric";
 import { useDialogContext } from "../layout/dialogs/DialogContext";
-import { Actions } from "../common/Actions";
-import { IIconButtonAction } from "../common/IconButtonWrapper";
+import { Actions } from "../common/actions/Actions";
+import { IIconButtonAction } from "../common/actions/IconButtonWrapper";
 import { getCommonActions } from "./getCommonActions";
 
 export const MetricHeaderActions: React.FC<{ metric: IMetric }> = ({
